Initialize app with data from sampledata/
Configure a "raw-loader" in webpack.config.js and typings.d.ts
so that .txt-files can be imported as strings.

Note that you need to run `npm i` after pulling to install raw-loader!

Import the two files in sampledata/ and use these as defaults in App.

Add a componentDidMount() hook in AppComponent that parses the default
strings into lists by calling parsePlanTemplateString() and
parsePersonsString().